Added new dashboard view file 	new file:   protected/modules/dashboard/views/dashboard/dashboard_main.php
This file is just a copy of a templated page and contains a lot of embedded psudeo code. However, I am committing it as the file will be used in other branches as well. I will spend some time adding functions to the dashboard, a lot of them copied from existing files (mostly profile edit).